fix(csm-tts): build depth-decoder causal mask in explicit fp32
Under the sglang model loader the default dtype is bf16, and
torch.finfo(torch.float32).min overflows bf16 in masked_fill_
(RuntimeError at CsmDepthDecoder init). Construct the additive mask
buffer in fp32; the use-site .to(hidden.dtype) cast keeps softmax
semantics identical (-3.4e38 rounds to -inf in bf16, exp()==0 either way).

+++ sglang_omni/models/csm_tts/modeling.py
@@ -364,7 +364,7 @@ class CsmDepthDecoder(nn.Module):
 
         # Additive causal mask over the full static window; row p hides every
         # key slot > p, which is also what makes per-frame KV reset a no-op.
-        causal = torch.zeros(ctx, ctx)
+        causal = torch.zeros(ctx, ctx, dtype=torch.float32)
         causal.masked_fill_(
             torch.ones(ctx, ctx, dtype=torch.bool).triu(1),
             torch.finfo(torch.float32).min,
